always run force activation, not matter WP_DEBUG
In all practical uses, we always want the force_active plugins
to be forced active, even when WP_DEBUG is true. What we really
need is a flag to allow us to turn on the force_deactive plugins
for dev purposes.

This always enables forced activation, and enables forced
deactivation unless WP_DEBUG is true.