override oembed HTML for lazyloading only on front-end requests
Since WP will only refresh the HTML for an oembed once per day,
we can't let it cache a non-lazyload version in the admin,
because then it will use that same version on the front end.

This stores the filtered HTML in an option, keyed by URL, so
that we can override the HTML only on front-end requests.

Updates https://central.tri.be/issues/64225