Add a changelog.md file
This changeset sets an initial version of 1.2. Three years ago there was the 1.1 version and a _lot_ has happened since then. My selection of 1.2 was just because it is an increase from the last stated version. If anyone feels strongly about calling the version something else, that works for me.

I also didn't attempt to summarize all of the changes in the past 3 years. Or even the past 3 months. I just hand selected a few token items to throw in the *Added* and the *Changed* sections. If we want to call out more things in those lists that are worthwhile, let's get them added.

Ultimately this is just to get us to a place where we can start tracking additions/changes/deletions and having awareness of what version of Square One that projects are running.

See: https://central.tri.be/issues/106088