Adding a base ansible directory that should be usable for most projects
This setup offloads all of the tasks to a separate repo: `moderntribe/tribe-ansible-roles` and the starter template focuses on variable declarations (with comments) and ansible-galaxy included roles, and playbooks (the things that are used to execute deploys). This new ansible setup has some speed improvements and grants us the ability to deploy server changes, database changes, and code changes independently.

Additionally, I've fixed the grunt install to instead be grunt-cli. And yarn is no longer installed via npm but instead as an apt package.

I threw in a pile-o comments to hopefully clear up some of the questions and I've bolstered the README in the ansible directory to help ease people into ansible awareness.