increase the delay before reprocessing incrementally
After the first failure, the default delay will be 10s, based
on the default priority. After each failure, the priority
increments, and the delay follows suit. After 100 failures,
we're looking at nearly 2 minutes between attempts.

The idea is that if it keeps failing, it's a more persistent
issue. Constantly re-trying won't help, and may even be
contributing to the problem.