Adding pre-requisite instructions for ansible
I fought with this a lot and came up short. Luckily @TaiSHiNet saved me. I'm adding this to the docs in hopes that this bits fewer people in the butt.

Note: Since this is a minor doc change, I'm thinking we don't need a version bump in changelog.md. Feel free to tell me I'm wrong and require a changelog addition, @speerface.