Sanitize SVG media uploads
The upload of SVG files was allowed, but not sanitized. Depending on the circumstances, SVG can be parsed with the browser's XML parser, thus executing Javascript, for instance. The vector of attack is very narrow though, because the file would have to been uploaded in the admin area and displayed inline or viewed directly in the front-end for the XSS to occur.

This PR uses the package "enshrined/svg-sanitize" from Packagist to sanitize SVGs through the "wp_handle_upload_prefilter" filter. The sanitization process was mostly taken from "Safe SVG" plugin, which is a WordPress plugin made by the same author of the packagist package.