feat: disable object cache when running tests
Our base test suite runs fine with object cache enabled. But once you
try to use it on multisite, we start to see conflicts between the
acceptance suite and the webdriver suite, with differing results
depending on the order the tests run.

By setting the environment variable DISABLE_OBJECT_CACHE, we prevent
the object cache drop-in from fully loading when tests are running.

Note that this is likely the first step towards moving our object
cache file out to tribe-libs, where we could maintain separate
packages for different caching backends.