Reorganization of Docs (#704)
* Rearange docs into a more logical structure

* Settle on Getting-Started restructure

* Move p2p doc to basics

* Convert doc links to full path

All links to other documents should be using the full path:
`/docs/{folder}/{file}.md`
This helps maintain valid links even if the file containing the link
is moved in the future. Also, if we restructure in the future, it's
easier to find a file reference by the full path rather than a mix of
full & relative paths.